# content/agent-api/openapi.yaml
openapi: 3.0.3
info:
  title: Agent API
  version: 0.1.0
  description: Production-grade API for autonomous onchain execution.
servers:
  - url: https://api.barterpayments.xyz
paths:
...
  /wallet/balance:
    post:
      summary: Get token balance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                chain:
                  type: string
                address:
                  type: string
                tokenAddress:
                  type: string
                decimals:
                  type: integer
              required: [chain, address, tokenAddress]
      responses:
        '200':
          description: Token balance
          content:
            application/json:
              schema:
                type: object
                properties:
                  address:
                    type: string
                  tokenAddress:
                    type: string
                  balance:
                    type: string
                  rawBalance:
                    type: string
                  decimals:
                    type: integer
...